Game Kind Of Done
A few issues remain, and haven't tested with actual places due to my
emulator crashing whenever I try to get the places from the server, so
I've commented out the code which gets places, for testing purposes. To
test the game with this code, go to the game tab and just click the map,
then visit.